koala-gains-backend/pyproject.toml
[tool.poetry]
name = "koala_gains"
version = "0.1.0"
description = ""
authors = ["DoDAO"]
readme = "README.md"

[virtualenvs]
create = true
in-project = true

[tool.poetry.dependencies]
python = "^3.10"
langgraph = "0.2.70"
langchain = "0.3.18"
langchain-core = "0.3.34"
langchain-community = "0.3.17"
langchain-openai = "0.3.4"
scrapingant-client = "^2.1.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
typing-extensions = "^4.12.2"
langchain-google-community = "2.0.4"
asyncio = "^3.4.3"
flask = "^3.1.0"
markdown = "^3.7"
reportlab = "^4.2.5"
boto3 = "^1.36.1"
beautifulsoup4 = "^4.12.3"
flask-cors = "^5.0.0"
linkedin-scraper = "^2.11.5"
langchain-groq = "^0.2.4"
flask-pydantic = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
